requirements.txt
urllib3~=1.25.8
PyQt5~=5.15.6
selenium~=3.141.0
Pillow~=9.0.1